+++ assets/js/app.js
@@ -21,7 +21,7 @@ import "phoenix_html";
 import { Socket } from "phoenix";
 import { LiveSocket } from "phoenix_live_view";
 import topbar from "../vendor/topbar";
-import { format, parseISO } from "date-fns";
+import { renderDateTimeInLocalTimezone } from "./localDateTime";
 
 let csrfToken = document
   .querySelector("meta[name='csrf-token']")
@@ -44,27 +44,4 @@ liveSocket.connect();
 // >> liveSocket.disableLatencySim()
 window.liveSocket = liveSocket;
 
-const items = document.querySelectorAll('[role~="local-datetime"]');
-items.forEach((item) => {
-  const date = parseISO(item.dataset.datetime);
-  const days = getDaysPassed(date);
-
-  const html =
-    getDaysPassed(date) > 0
-      ? format(date, "yyyy-MM-dd HH:mm:ss")
-      : format(date, "HH:mm:ss");
-
-  item.innerHTML = html;
-});
-
-function getDaysPassed(startDate) {
-  const currentDate = new Date();
-
-  // Calculate the time difference in milliseconds
-  const timeDifference = currentDate - startDate;
-
-  // Convert milliseconds to days
-  const daysPassed = Math.floor(timeDifference / (1000 * 60 * 60 * 24));
-
-  return daysPassed;
-}
+renderDateTimeInLocalTimezone();
